requirements.txt
requests>=2.31.0
pandas>=2.2.0
SQLAlchemy>=2.0.21
PyMySQL>=1.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
datadog>=0.52.1
ddtrace>=3.14.1
psutil>=5.9.5